move test data generation into separate package
